rustc_driver_impl::signal_handler
Macro raw_errln

macro raw_errln($tokens:tt) {
    ...
}

We don’t really care how many bytes we actually get out. SIGSEGV comes for our head. Splash stderr with letters of our own blood to warn our friends about the monster.